requirements.txt
fastapi
uvicorn
pydantic
PyPDF2
python-multipart
pinecone-client
nltk
python-dotenv
transformers
sentence-transformers